[TWINFACES-157] deleted file api-service for twin links
