+++ src/entities/twin/api/api-service.ts
@@ -99,12 +99,29 @@ export function createTwinApi(settings: ApiSettings) {
     });
   }
 
+  function getLinks({ twinId }: { twinId: string }) {
+    return settings.client.GET("/private/twin/{twinId}/v2", {
+      params: {
+        header: getApiDomainHeaders(settings),
+        path: { twinId: twinId },
+        query: {
+          showTwin2TwinLinkMode: "DETAILED",
+          showTwinLink2UserMode: "DETAILED",
+          showTwinLink2LinkMode: "DETAILED",
+          showTwinField2TwinMode: "DETAILED",
+          showTwinByLinkMode: "GREEN",
+        },
+      },
+    });
+  }
+
   return {
     search,
     getById,
     update,
     getFieldById,
     getHistory,
+    getLinks,
   };
 }
