+++ src/entities/twinLink/api/index.ts
@@ -1,2 +1,2 @@
-export * from "./api-service";
 export * from "./types";
+export * from "./hooks";
