+++ src/features/api-context-provider/api-context-provider.tsx
@@ -54,7 +54,6 @@ import {
   createFactoryPipelineApi,
   FactoryPipelineApi,
 } from "@/entities/factoryPipeline";
-import { createTwinLinksApi, TwinLinkApi } from "@/entities/twinLink";
 
 export interface ApiContextProps {
   domain: DomainApi;
@@ -78,7 +77,6 @@ export interface ApiContextProps {
   factory: FactoryApi;
   factoryPipeline: FactoryPipelineApi;
   spaceRole: PermissionSpaceRoleApi;
-  twinLink: TwinLinkApi;
 }
 
 export function ApiContextProvider({
@@ -119,7 +117,6 @@ export function ApiContextProvider({
         factory: createFactoryApi(settings),
         factoryPipeline: createFactoryPipelineApi(settings),
         spaceRole: createPermissionSpaceRoleApi(settings),
-        twinLink: createTwinLinksApi(settings),
       }}
     >
       {children}
